move ui utils to their own file
