+++ packages/cli/lib/links.js
@@ -1,4 +1,3 @@
-const supportsHyperlinks = require('supports-hyperlinks');
 const { getEnv } = require('@hubspot/cli-lib/lib/config');
 const { ENVIRONMENTS } = require('@hubspot/cli-lib/lib/constants');
 const { getHubSpotWebsiteOrigin } = require('@hubspot/cli-lib/lib/urls');
@@ -121,18 +120,9 @@ const openLink = (accountId, shortcut) => {
   logger.success(`We opened ${match.url} in your browser`);
 };
 
-const link = (linkText, url, options = {}) => {
-  if (supportsHyperlinks.stdout) {
-    return ['\u001B]8;;', url, '\u0007', linkText, '\u001B]8;;\u0007'].join('');
-  } else {
-    return options.fallback ? `${linkText}: ${url}` : linkText;
-  }
-};
-
 module.exports = {
   getSiteLinks,
   getSiteLinksAsArray,
   logSiteLinks,
   openLink,
-  link,
 };
